requirements.txt
numpy
pandas
requests
python-dotenv
alpaca-py 